+++ qbe_backend/src/lower_lang.rs
@@ -1,0 +1,79 @@
+pub enum LowerLang {
+    CompUnit(String, Vec<Decl>),
+}
+
+pub type DataField = (Type, String);
+
+pub enum Decl {
+    Data(String, Vec<DataField>),
+    FuncDecl(String, Option<Linkage>, Type, FuncArgs, Vec<Stmt>),
+}
+
+pub enum FuncArgs {
+    Fixed(Vec<FuncArg>),
+    Variadic(FuncArg, Vec<FuncArg>),
+}
+
+pub enum FuncArg {
+    Named(String, Type),
+}
+
+pub enum Stmt {
+    FuncCall(String, Vec<Value>),
+    Jmp(String),
+    Jnz(Expr, String, String),
+    Lbl(String),
+    Ret(Value),
+    VarDecl(String, Scope, Expr),
+}
+
+pub enum Expr {
+    Value(Value),
+    FuncCall(String, Type, Vec<Value>),
+}
+
+pub enum Value {
+    ConstW(i32),
+    VarRef(String, Type, Scope),
+}
+
+pub enum Linkage {
+    Export,
+}
+
+#[derive(Clone, Copy)]
+pub enum Type {
+    B,
+    D,
+    H,
+    L,
+    S,
+    W,
+}
+
+pub enum Scope {
+    Func,
+    Global,
+}
+
+pub trait Typed {
+    fn r#type(&self) -> Type;
+}
+
+impl Typed for Expr {
+    fn r#type(&self) -> Type {
+        match self {
+            Expr::Value(value) => value.r#type(),
+            Expr::FuncCall(_, r#type, _) => *r#type,
+        }
+    }
+}
+
+impl Typed for Value {
+    fn r#type(&self) -> Type {
+        match self {
+            Value::ConstW(_) => Type::W,
+            Value::VarRef(_, r#type, _) => *r#type,
+        }
+    }
+}
